+++ bsp/posix/board/board.c
@@ -3,9 +3,8 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-#ifndef _BOARD_H_
-#define _BOARD_H_
 
-#include "memory_map.h"
-
-#endif /* _BOARD_H_ */
+void board_init(void)
+{
+    // Intentionally empty
+}