+++ src/hal/posix/init.c
@@ -1,0 +1,15 @@
+/**
+ * Copyright (c) 2021 Wavious LLC.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+void hal_timer_init(void)
+{
+    // Intentionally empty
+}
+
+void hal_subsystem_init(void)
+{
+    // Intentionally empty
+}